slider with minor bug
