+++ imports/ui/Home.jsx
@@ -12,7 +12,6 @@ import Link from '@material-ui/core/Link';
 import useStyles from './style/NavStyle';
 import UserList from "./UserList/UserList.jsx";
 import {Meteor} from "meteor/meteor";
-import SliderShow from "./SliderShow"
 class home extends Component {
 
     // constructor(props) {
@@ -71,9 +70,8 @@ class home extends Component {
                 <link rel="stylesheet"
                       href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"/>
                 <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons"/>
-                <SliderShow/>
-                <Categories/>
                 <List/>
+                <Categories/>
             </React.Fragment>
         );
     }
