+++ imports/ui/postProcedures/PostDetailUnderOneCategory.jsx
@@ -10,6 +10,7 @@ class PostDetailUnderOneCategory extends React.Component {
     }
     conditionalRender(){
         if (this.props.unposted.category !== "") {
+            console.log("render item detail panel");
             return (<ItemDetail/>);
         }
     }
