+++ imports/ui/postProcedures/Review.jsx
@@ -17,7 +17,11 @@ import { resetPost,closePostReview } from '../../actions';
 class Popup extends React.Component {
     constructor(props) {
         super(props);
+        this.state = {
+          isDone: false
+        }
         this.handleClick = this.handleClick.bind(this);
+        this.resetAndclose = this.resetAndclose.bind(this);
     }
     handleClick(){
       console.log(Meteor.user());
@@ -43,14 +47,40 @@ class Popup extends React.Component {
             Meteor.call("createItem",withOwnerInfo_Item);
         });
       }
-        alert('Add done!');
-        this.props.reset();
-        this.props.close();
+        this.setState({
+          isDone:true
+        })
+    }
+    resetAndclose(){
+      this.setState({
+        isDone:false
+      })
+      this.props.close();
+      this.props.reset();
     }
     render() {
       return (
-        <div className='popup'>
+          <div className='popup'>
           <div className='popup_inner'>
+          {this.state.isDone
+            ? <div>
+                <br/>
+                <br/>
+                <br/>
+
+                <h3>done!</h3>
+
+                <svg version="1.1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 130.2 130.2">
+                  <circle class="path_circle" fill="none" stroke="#73AF55" stroke-width="6" stroke-miterlimit="10" cx="65.1" cy="65.1" r="62.1"/>
+                  <polyline class="path_check" fill="none" stroke="#73AF55" stroke-width="6" stroke-linecap="round" stroke-miterlimit="10" points="100.2,40.2 51.5,88.8 29.8,67.5 "/>
+                </svg>
+
+                <br/>
+                <br/>
+                <br/>
+                <button type = 'close' onClick = {this.resetAndclose}>Close</button>
+              </div>
+            : <div>
                 <h2>Your item info :</h2>
                 <br/>
                 <div>Title: {this.props.detail.title}</div>
@@ -73,7 +103,9 @@ class Popup extends React.Component {
                 <br/>
                 <button type = 'close' onClick={this.props.close}>Need Revise</button>
                 <button type = 'submit' onClick={this.handleClick}>Submit</button>
-          </div>
+              </div>
+          }
+          </div>        
         </div>
       );
     }
